+++ core/abc/entities/menus/menu_entry.rb
