+++ core/abc/entities/menus/menuable.rb
