+++ core/abc/entities/menus/path.rb
